Add support for iterable dataset loading
Passing a streaming dataset is now possible, with the following
differences:

1. Examples must be padded to max sequence length
2. max_training_steps (and optionally max_evaluation_steps)
   must be set in the TrainArguments

The reason for 1. is that the collate_fn used for the normal
(non-streaming) dataset passed is suitable only for batched
hf/torch datasets. Since in the iterable case the batching is
handled by tf datasets, existing collate_fn's cannot be used.

This commit adapts the train and eval loops to request only the exact
amount of batches as expected, to prevent tf core warning
Local rendezvous is aborting with status: OUT_OF_RANGE: End of sequence